// staterkit/src/types/common.ts
export interface Select {
  value: string | number
  label: string
  profile?: string
  data?: Select[]
}